Moves export_test.go artifacts to package_test.go, which becomes local to the instancepoller package.
Adds mock generation directive for link-layer devices and addresses
indirections.
