--- apiserver/facades/controller/instancepoller/export_test.go
@@ -1,16 +1,0 @@
-// Copyright 2015 Canonical Ltd.
-// Licensed under the AGPLv3, see LICENCE file for details.
-
-package instancepoller
-
-import "github.com/juju/juju/state"
-
-type Patcher interface {
-	PatchValue(ptr, value interface{})
-}
-
-func PatchState(p Patcher, st StateInterface) {
-	p.PatchValue(&getState, func(*state.State, *state.Model) StateInterface {
-		return st
-	})
-}
